+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Log messages start with upper case
 - Update Readme
 - execute list repos, list organizations and create ticket always in kidney workspace
+- Error if ticket already exists
 
 ### Removed